refactor: logging.js
Now with unit tests!
