+++ server.js
@@ -9,7 +9,7 @@ const ejs = require('ejs')
 // Classes
 const server = require('./web/js/classes');
 const rmUtils = require('./rmUtilities');
-const logging = require('./logging');
+const logging = require('./src/rm/logging');
 const handlers = require("./src/rm/handlers");
 const transmit = require('./src/rm/transmit');
